+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2025-07-10
+
+### Added
+
+-   New `clean` command to remove models.
+    -   `clean configs`: Removes only the model configurations created by ollamaX (those with a `-ctx...k` suffix).
+    -   `clean all`: Removes all Ollama models from the system after a confirmation prompt.
+
 ## [1.1.0] - 2025-07-10
 
 ### Changed